// src/app/Model/persona.model.ts
export class persona {
    nombre: String = "";
    primape: String = "";
    segape: String = "";
    pais: String = "";
    estado: String = "";
    claveEntidad: number = 0;
    municipio: String = "";
    origen: String = "";
    nacionalidad: String = "";
    sexo: String = "";
    fecha_nac: String = "";
    visto_ultima: String = "";
    autoridad: String = "";
    coordenadaX: number = 0;
    coordenadaY: number = 0;
}